+++ package.json
@@ -47,10 +47,8 @@
     "@types/qrcode": "^1.4.1",
     "@types/uuid": "^8.3.1",
     "@types/validator": "^13.6.3",
-    "@typescript-eslint/eslint-plugin": "^6.9.0",
-    "@typescript-eslint/parser": "^6.9.0",
-    "ajv": "^8.12.0",
-    "ajv-formats": "^2.1.1",
+    "@typescript-eslint/eslint-plugin": "^4.28.5",
+    "@typescript-eslint/parser": "^4.28.5",
     "cbor": "^7.0.6",
     "commitizen": "^4.2.5",
     "eslint": "^7.32.0",
@@ -65,7 +63,7 @@
     "prettier": "^2.3.2",
     "qrcode": "^1.5.1",
     "quicktype": "^15.0.260",
-    "rollup": "^2.79.1",
+    "rollup": "^2.56.2",
     "rollup-plugin-commonjs": "^10.1.0",
     "semantic-release": "^17.4.4",
     "ts-jest": "^29.1.1",
@@ -74,6 +72,8 @@
   },
   "dependencies": {
     "@govtechsg/jsonld": "^0.1.0",
+    "ajv": "^8.6.2",
+    "ajv-formats": "^2.1.0",
     "cross-fetch": "^3.1.5",
     "debug": "^4.3.2",
     "ethers": "^5.7.2",
