+++ src/shared/validate/validate.ts
@@ -1,10 +1,9 @@
+import { ErrorObject, ValidateFunction } from "ajv";
 import { getLogger } from "../logger";
 import { SchemaId } from "../@types/document";
 // don't change this otherwise there is a cycle
 import { getData } from "../utils/utils";
 import { Kind } from "../utils/@types/diagnose";
-import { ValidateFunction } from "ajv/dist/core";
-import { ErrorObject } from "ajv";
 const logger = getLogger("validate");
 
 export const validateSchema = (document: any, validator: ValidateFunction, kind?: Kind): ErrorObject[] => {
